Add copyright header and remove unused interface
