--- src/main/java/io/github/eb4j/ebview/gui/IMainMenu.java
@@ -1,6 +1,0 @@
-package io.github.eb4j.ebview.gui;
-
-public interface IMainMenu {
-
-    void invokeAction(String action, int modifiers);
-}
