+++ src/main/java/io/github/eb4j/ebview/gui/MainWindowMenu.java
@@ -1,9 +1,26 @@
+/*
+ * EBViewer, a dictionary viewer application.
+ * Copyright (C) 2022 Hiroshi Miura.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 package io.github.eb4j.ebview.gui;
 
 import io.github.eb4j.ebview.EBViewer;
 import io.github.eb4j.ebview.dictionary.DictionariesManager;
 import io.github.eb4j.ebview.gui.dialogs.AboutDialog;
-import io.github.eb4j.ebview.gui.preferences.IPreferencesController;
 import io.github.eb4j.ebview.gui.preferences.PreferenceController;
 import io.github.eb4j.ebview.utils.LStrings;
 
@@ -28,7 +45,7 @@ import java.lang.reflect.Method;
 
 import static io.github.eb4j.ebview.utils.ResourceUtil.APP_ICON_32X32;
 
-public class MainWindowMenu implements ActionListener, MenuListener, IMainMenu {
+public class MainWindowMenu implements ActionListener, MenuListener {
 
     private final JFrame app;
     private final DictionariesManager manager;
