+++ src/main/java/io/github/eb4j/ebview/utils/Stemmer.java
@@ -1,3 +1,24 @@
+/*
+ * EBViewer, a dictionary viewer application.
+ * Copyright (C) 2022 Hiroshi Miura.
+ * Copyright (C) 2000-2006 Keith Godfrey, Maxym Mykhalchuk, and Henry Pijffers
+ *              2007 Didier Briel, Zoltan Bartko
+ *              2008 Alex Buloichik
+ *              2015 Didier Briel, Aaron Madlon-Kay*
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 package io.github.eb4j.ebview.utils;
 
 import org.apache.commons.lang3.StringUtils;
